chore: move remark plugins to test fixture
